OpenMPI and Coarray Fortran on AWS
Note the MCA configuration.  Without this additional line, one gets
errors of this form when trying to run MPI codes:
```
Open MPI detected an inbound MPI TCP connection request from a peer
that appears to be part of this MPI job (i.e., it identified itself as
part of this Open MPI job), but it is from an IP address that is
unexpected.  This is highly unusual.

The inbound connection has been dropped, and the peer should simply
try again with a different IP interface (i.e., the job should
hopefully be able to continue).

  Local host:          ip-172-31-26-140
  Local PID:           104
  Peer hostname:       ip-172-31-26-140 ([[12,0],2])
  Source IP of socket: 172.31.26.140
  Known IPs of peer:
```